# packages/orchestrator/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Linting Orchestrator API",
    "version": "0.13.1",
    "description": "The linting orchestrator uses Spectral and custom rule engines.\nIt provides a REST API for uploading documents, submitting lint jobs, and retrieving results.\n\n**Deployment Modes:**\n- **Light Mode** (Embedded): CLI embeds server, zero-config, perfect for CI/CD\n- **Standalone Mode**: Independent server process, long-running, multi-user\n- **Companion Mode**: Runs alongside the MCP OpenAPI Analyzer for integrated document lifecycle",
    "contact": {
      "name": "Cisco DevNet"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response returned by all endpoints on failure",
        "required": [
          "error",
          "timestamp"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error category (e.g., Not Found, Validation Error, Internal Server Error)",
            "example": "Not Found"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message with context",
            "example": "Document not found: abc-123"
          },
          "details": {
            "type": "object",
            "description": "Additional error details such as validation errors or field-level issues",
            "additionalProperties": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp when the error occurred",
            "example": "2026-03-05T12:00:00.000Z"
          }
        }
      },
      "JobStatus": {
        "type": "string",
        "description": "Lifecycle status of a lint job",
        "enum": [
          "queued",
          "running",
          "completed",
          "completed_with_errors",
          "failed",
          "timeout",
          "cancelled"
        ],
        "example": "completed"
      },
...
      "ResultSummary": {
        "type": "object",
        "description": "Aggregated issue counts by severity level",
        "properties": {
          "totalIssues": {
            "type": "integer",
            "description": "Total number of issues found across all severity levels",
            "example": 13
          },
          "errorCount": {
            "type": "integer",
            "description": "Number of issues with error severity (must fix)",
            "example": 5
          },
          "warningCount": {
            "type": "integer",
            "description": "Number of issues with warning severity (should fix)",
            "example": 8
          },
          "infoCount": {
            "type": "integer",
            "description": "Number of informational issues",
            "example": 0
          },
          "hintCount": {
            "type": "integer",
            "description": "Number of hint-level suggestions",
            "example": 0
          }
        }
      },
...
      "JobSummary": {
        "type": "object",
        "description": "Lightweight summary of a lint job for list views",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid",
            "description": "Unique job identifier",
            "example": "94c9db52-92fc-4733-806a-fe6d206109aa"
          },
          "documentId": {
            "type": "string",
            "description": "Identifier of the linted document",
            "example": "41c1d4aa-5825-4313-8ae5-737490707b5b"
          },
          "rulesetName": {
            "type": "string",
            "description": "Name of the ruleset used for linting",
            "example": "pubhub"
          },
          "rulesetVersion": {
            "type": "string",
            "description": "Version of the ruleset used for linting",
            "example": "1.1.0"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp when the job completed",
            "example": "2026-03-05T12:00:01.000Z"
          },
          "totalExecutionTime": {
            "type": "number",
            "description": "Total execution time in milliseconds",
            "example": 321
          },
          "summary": {
            "$ref": "#/components/schemas/ResultSummary"
          }
        }
      },
...
      "JobListResponse": {
        "type": "object",
        "description": "Paginated list of lightweight job summaries",
        "properties": {
          "jobs": {
            "type": "array",
            "description": "Array of job summary objects",
            "items": {
              "$ref": "#/components/schemas/JobSummary"
            }
          },
          "pagination": {
            "type": "object",
            "description": "Pagination metadata",
            "properties": {
              "total": {
                "type": "integer",
                "description": "Total number of jobs matching the filter criteria",
                "example": 10
              },
              "limit": {
                "type": "integer",
                "description": "Maximum number of results returned in this page",
                "example": 50
              },
              "offset": {
                "type": "integer",
                "description": "Number of results skipped for pagination",
                "example": 0
              },
              "hasMore": {
                "type": "boolean",
                "description": "Whether more results are available beyond this page",
                "example": false
              }
            }
          },
          "sessionId": {
            "type": "string",
            "description": "Runtime session ID used to filter stale jobs from previous server instances",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/lint/jobs": {
      "get": {
        "operationId": "listJobs",
        "summary": "List jobs (lightweight)",
        "tags": [
          "Jobs"
        ],
        "description": "List lint jobs with optional filtering. Returns lightweight job summaries (document ID only, no metadata).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "status",
            "required": false,
            "description": "Filter by job status. Comma-separated for multiple values."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "documentId",
            "required": false,
            "description": "Filter by document ID"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "rulesetName",
            "required": false,
            "description": "Filter by ruleset name"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "startDate",
            "required": false,
            "description": "Filter jobs created after this date (ISO 8601)"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "in": "query",
            "name": "endDate",
            "required": false,
            "description": "Filter jobs created before this date (ISO 8601)"
          },
          {
            "schema": {
              "type": "integer",
              "default": 50
            },
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Maximum number of results"
          },
          {
            "schema": {
              "type": "integer",
              "default": 0
            },
            "in": "query",
            "name": "offset",
            "required": false,
            "description": "Pagination offset"
          },
          {
            "schema": {
              "type": "string",
              "default": "timestamp"
            },
            "in": "query",
            "name": "sortBy",
            "required": false,
            "description": "Field to sort by"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "in": "query",
            "name": "sortOrder",
            "required": false,
            "description": "Sort direction"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3099",
      "description": "Local development server (default port)"
    }
  ],
...
}